pyproject.toml
[tool.poetry]
name = "wordlinator"
version = "0.1.0"
description = "Bot to pull WordleGolf scores from twitter and record in GSheets"
authors = ["Brad Brown <brad@bradsbrown.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
google-api-python-client = "^2.49.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.1"
rich = "^12.4.4"
python-dateutil = "^2.8.2"
httpx = "^0.23.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
mypy = "^0.960"
flake8 = "^4.0.1"
isort = "^5.10.1"

[tool.poetry.scripts]
wordlinator = "wordlinator.app:sync_main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
